requirements.txt
langchain==0.0.293
PyPDF2==3.0.1
python_docx==0.8.11
sentence_transformers==2.2.2
streamlit
streamlit_chat==0.1.1

openai
faiss-cpu
